Move terminalActions, electron-browser->browser
Part of #69115
